requirements.txt
Django==1.7.1
PyYAML==3.10
Twisted==13.2.0
altgraph==0.10.2
argparse==1.2.2
bdist-mpkg==0.5.0
beautifulsoup4==4.3.1
bonjour-py==0.3
dateutils==0.6.6
dj-database-url==0.3.0
gunicorn==19.1.1
httplib2==0.8
macholib==1.5.1
matplotlib==1.3.1
modulegraph==0.10.4
nltk==2.0.1rc1
numpy==1.8.0rc1
oauth-python-twitter==1.0
oauthlib==0.7.2
py2app==0.7.3
pyOpenSSL==0.13.1
pyobjc-core==2.5.1
pyobjc-framework-Accounts==2.5.1
pyobjc-framework-AddressBook==2.5.1
pyobjc-framework-AppleScriptKit==2.5.1
pyobjc-framework-AppleScriptObjC==2.5.1
pyobjc-framework-Automator==2.5.1
pyobjc-framework-CFNetwork==2.5.1
pyobjc-framework-Cocoa==2.5.1
pyobjc-framework-Collaboration==2.5.1
pyobjc-framework-CoreData==2.5.1
pyobjc-framework-CoreLocation==2.5.1
pyobjc-framework-CoreText==2.5.1
pyobjc-framework-DictionaryServices==2.5.1
pyobjc-framework-EventKit==2.5.1
pyobjc-framework-ExceptionHandling==2.5.1
pyobjc-framework-FSEvents==2.5.1
pyobjc-framework-InputMethodKit==2.5.1
pyobjc-framework-InstallerPlugins==2.5.1
pyobjc-framework-InstantMessage==2.5.1
pyobjc-framework-LatentSemanticMapping==2.5.1
pyobjc-framework-LaunchServices==2.5.1
pyobjc-framework-Message==2.5.1
pyobjc-framework-OpenDirectory==2.5.1
pyobjc-framework-PreferencePanes==2.5.1
pyobjc-framework-PubSub==2.5.1
pyobjc-framework-QTKit==2.5.1
pyobjc-framework-Quartz==2.5.1
pyobjc-framework-ScreenSaver==2.5.1
pyobjc-framework-ScriptingBridge==2.5.1
pyobjc-framework-SearchKit==2.5.1
pyobjc-framework-ServiceManagement==2.5.1
pyobjc-framework-Social==2.5.1
pyobjc-framework-SyncServices==2.5.1
pyobjc-framework-SystemConfiguration==2.5.1
pyobjc-framework-WebKit==2.5.1
pyparsing==2.0.1
python-dateutil==1.5
pytz==2013.7
requests==2.4.3
requests-oauthlib==0.4.2
scipy==0.13.0b1
simplejson==3.6.5
six==1.4.1
twilio==3.6.4
twitter==1.15.0
twython==3.2.0
unittest2==0.5.1
virtualenv==1.11.6
wsgiref==0.1.2
xattr==0.6.4
zope.interface==4.1.1
